feat: enable theme from query
