+++ deno/pages/lang.ts
@@ -1,0 +1,1 @@
+export const fr = ["/allô"];
